+++ impluse-hfs/ImpTextEncodingConverter.m
@@ -89,7 +89,7 @@
 		return nil;
 	}
 
-	S(outputBuf[0], actualOutputLengthInBytes / sizeof(UniChar));
+	S(outputBuf[0], (u_int16_t)(actualOutputLengthInBytes / sizeof(UniChar)));
 
 	return unicodeData;
 }
